api/requirements.txt
fastapi
uvicorn[standard]
rq
redis
ocrmypdf
python-multipart
Pillow
